chore: OpenCV input data augmentation refactor
